+++ src/renderer/views/wallet/PortfolioView/utils.ts
@@ -1,0 +1,4 @@
+export enum PortfolioTabKey {
+  ChartView,
+  CardView
+}